Increase loading screen FPS
Now that it's rendered in a background thread, we can have a smooth loading bar at virtually no performance cost.

+++ apps/openmw/mwgui/loadingscreen.hpp
@@ -50,8 +50,8 @@ namespace MWGui
         const VFS::Manager* mVFS;
         osg::ref_ptr<osgViewer::Viewer> mViewer;
 
-        unsigned long mLastWallpaperChangeTime;
-        unsigned long mLastRenderTime;
+        double mLastWallpaperChangeTime;
+        double mLastRenderTime;
         osg::Timer mTimer;
 
         size_t mProgress;
